re-run go-generate to fix up LICENSE text (#2171)
I ran a script that added the license to all go files, but forgot that it would operate on generated files which would of course be overwritten. This change adds the license to generated code where possible, and removes it from generated mocks. Aside from changes to templates, this is all fresh output from make go-generate

fbshipit-source-id: 5fae422